Refactor stage contract types by concept
